+++ api/db_models/__init__.py
@@ -1,9 +1,13 @@
-from .models import User, Address, MenuItem, Menu, OrderItem, Order,\
-PaymentMethod, Payment, TransactionItem, Transaction, ReservationItem,\
-Reservation, Restaurant, ShipmentMethod, Shipment, Invoice, InvoiceItem,\
-Event, EventItem, Information
+
+from .models import (Address, Cart, CartItem, Event, EventItem, Information,
+                     Invoice, InvoiceItem, Menu, MenuItem, Order, OrderItem,
+                     Payment, PaymentMethod, Reservation,
+                     Restaurant, Shipment, ShipmentMethod, Transaction,
+                     User, VerificationToken)
 
 __all__ = ['User', 'Address', 'MenuItem', 'Menu', 'OrderItem', 'Order', 'PaymentMethod',
-           'Payment', 'TransactionItem', 'Transaction', 'ReservationItem', 
+           'Payment', 'Transaction',
            'Reservation', 'Restaurant', 'ShipmentMethod', 'Shipment', 'Invoice',
-           'InvoiceItem', 'Event', 'EventItem', 'Information']
+
+           'InvoiceItem', 'Event', 'EventItem', 'Information',
+           'Cart', 'CartItem', 'VerificationToken']