+++ app.py
@@ -1,26 +1,31 @@
 """application entry point"""
-
+from os import getenv
+from flask import Flask
+from dotenv import load_dotenv
 from flask import Flask
 from flask_cors import CORS
+from flask_jwt_extended import JWTManager
 
-# models
 from api.core.base import Db as db
 from api.routes import api
 
+# load environment variables from .env file
+load_dotenv()
 
 # Initializing flask app
 app = Flask(__name__)
-app.secret_key = '8445a9af69bccd13de1a10de1de88158'
+app.secret_key = getenv('SECRET_KEY')
 
+# Setup the Flask-JWT-Extended extension
+app.config['JWT_SECRET_KEY'] = getenv('SECRET_KEY')
+jwt = JWTManager(app)
 
 # Initializing CORS for cross origin requests
 CORS(app, resources={r'/*': {'origins': '*'}}, supports_credentials=True)
 
-# models
 
 # initializing database with flask app
 db.initialize_app(app)
-
 
 # Route for seeing a data
 app.register_blueprint(api)