+++ client/src/pages/ProductsPage.js
@@ -1,17 +1,42 @@
 import { Helmet } from 'react-helmet-async';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 // @mui
 import { Container, Stack, Typography } from '@mui/material';
+import LoadingSpinner from '../sections/loadingSpinner/LoadingSpinner';
 // components
 import { ProductSort, ProductList, ProductFilterSidebar } from '../sections/products';
 // mock
 // we should provide an array of product categories
 import PRODUCTS from '../_mock/products';
-import ProductCartPopover from '../sections/products/ProductCartPopover';
+
 
 // ----------------------------------------------------------------------
 
+
 export default function ProductsPage() {
+
+  // get the data 
+  const api = `${process.env.REACT_APP_API}/dashboard/menus`;
+  console.log(api);
+  // menus state
+  const [menus, setMenus] = useState();
+  // request options
+  const requestOptions = {
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+  };
+
+  useEffect(() => {
+    // fetch the data
+    fetch(api, requestOptions).then((response) => response.json())
+      .then((data) => {
+        console.log(data.menus);
+        setMenus(data.menus);
+      }).catch((error) => {
+        console.log(error);
+      });
+  }, []);
+  sessionStorage.setItem('signup', 'true')
   const [openFilter, setOpenFilter] = useState(false);
 
   const handleOpenFilter = () => {
@@ -27,50 +52,37 @@ export default function ProductsPage() {
       <Helmet>
         <title> Your Cravings | Open Restaurant </title>
       </Helmet>
+      {
+        menus ?
+          <Container>
 
-      <Container>
-        <Typography variant="h4" sx={{ mb: 5 }}>
-          Your Cravings Solved
-        </Typography>
-
-        <Stack direction="row" flexWrap="wrap-reverse" alignItems="center" justifyContent="flex-end" sx={{ mb: 5 }}>
-          <Stack direction="row" spacing={1} flexShrink={0} sx={{ my: 1 }}>
-            <ProductFilterSidebar
-              openFilter={openFilter}
-              onOpenFilter={handleOpenFilter}
-              onCloseFilter={handleCloseFilter}
-            />
-            <ProductSort />
-          </Stack>
-        </Stack>
-
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5 , mt: 5}}>
-          Coco Tails
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
-        Take Away
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
-        Barques 
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
-        Burgers
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
-        Cakes
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
-        Ice cream
-        </Typography>
-        <ProductList products={PRODUCTS} />
-        <ProductCartPopover />
-      </Container>
+            <Typography variant="h4" sx={{ mb: 5 }}>
+              Your Cravings Solved
+            </Typography>
+            <Stack direction="row" flexWrap="wrap-reverse" alignItems="center" justifyContent="flex-end" sx={{ mb: 5 }}>
+              <Stack direction="row" spacing={1} flexShrink={0} sx={{ my: 1 }}>
+                <ProductFilterSidebar
+                  openFilter={openFilter}
+                  onOpenFilter={handleOpenFilter}
+                  onCloseFilter={handleCloseFilter}
+                />
+                <ProductSort />
+              </Stack>
+            </Stack>
+            {
+              menus.map((menu) => (
+                <>
+                  <Typography variant="h4" sx={{ mb: 5, mt: 5 }}>
+                    {menu.name}
+                  </Typography>
+                  <ProductList products={menu.items} />
+                </>
+              ))
+            }
+          </Container>
+          :
+          <LoadingSpinner />
+      }
     </>
   );
 }