+++ client/src/sections/reservation/new/PaymentForm.js
@@ -8,7 +8,7 @@ import Checkbox from '@mui/material/Checkbox';
 export default function PaymentForm() {
   return (
     <>
-      <Typography variant="h6" gutterBottom>
+      <Typography variant="h6" gutterBottom sx={{ my: 10 }}>
         Payment method
       </Typography>
       <Grid container spacing={3}>